lib/workload/stateless/stacks/case-manager/pyproject.toml
[project]
name = "case-manager"
version = "0.0.1"
description = "OrcaBus Case Manager"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-xray-sdk",
    "django==5.1.2",
    "djangorestframework==3.15.2",
    "django-cors-headers==4.5.0",
    "django-environ==0.11.2",
    "djangorestframework-camel-case==1.4.2",
    "drf-spectacular==0.27.2",

    # See psycopg[binary] or psycopg[c] impl https://www.psycopg.org/psycopg3/docs/basic/install.html
    "psycopg[binary]==3.2.3",
    "Werkzeug==3.0.6",
    "libumccr==0.4.0",
    "cachetools==5.5.0",
    "serverless-wsgi==3.0.5",
    # six and regex required by automatically generated EventBridge code binding
    # for sequencerunstatechange package
    "six==1.16.0",
    "regex==2024.9.11",
    "ulid-py==1.1.0",
]